v1.2
Improvements for the UI added menu, added font size menu for the edit
area
Added saving and opening for modpe .js, .txt and .coffee files. Added
improvements to the Console, A back/hide button and a clear button.
Added about window with links to the minecraft forums thread and github
page.